requirements.txt
Django	== 1.11.3
Pillow	== 4.2.1
django-countries ==	4.6.1
django-tinymce	== 2.6.0
